Move Logger implemetation under the Contexts folder
